Skip style check if app fails schema validation
If an app is invalid, it could be a minor thing, like missing a noun, or it could be a major thing, like defining a piece of the app that should be an object as a string. We do not have a way to tell how severe it is, so we should opt to not do the style check at all.