zapier convert: Fix auth tests
Most of these tests failed because I added heading and trailing spaces
to the trigger template. This commit fixes them by removing the heading
and trailing spaces with .trim() method before making assertion in the
test code. This is temporary because I think we can use a tool like
Prettier to format the generated code automitcally, so we don't have to
manually format it in the templates, which is such a pain.