Don't hang on calls to utils.printData
When running utils.printData on a headless terminal (such as a CI)
or a "skinny" terminal window, the call would hang indefinitely.
The call was getting into a never-ending while-loop because the
calculated window space was too small for output. Instead of going
into that while-loop, if the calculated window space is too small,
the printData call will switch from row output to plain output.